Move dictionary_product to util.py
